pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ocflindex"
version = "0.0.1"
dependencies = [
    "grpcio",
    "protobuf",
]

[tool.setuptools.packages.find]
where = ["ocflindex"]